requirements.txt
# Core
python-dotenv>=1.0.1
omegaconf>=2.3.0
pyyaml>=6.0.2
pandas>=2.2.2
numpy>=1.26.4

# FAISS (CPU by default; switch to GPU via conda if desired)
faiss-cpu>=1.8.0

# If the Vast.ai image has CUDA-enabled FAISS preinstalled, you can remove faiss-cpu above.
# For conda GPU install: conda install -c pytorch faiss-gpu

# Embeddings / Reranking
sentence-transformers>=3.0.1
FlagEmbedding>=1.2.10

# Generation (pick one provider path; keep both clients installed)
huggingface_hub>=0.24.5
transformers>=4.43.3
accelerate>=0.33.0
httpx>=0.27.0

# PDF / text parsing
pymupdf>=1.24.7

# Metrics / plots
scikit-learn>=1.5.1
matplotlib>=3.8.4

# Utilities
tqdm>=4.66.4
rich>=13.7.1

torch
pyarrow>=16.0.0
fastparquet>=2024.2.0
python-dotenv>=1.0.0
